+++ app/Exceptions/Handler.php
@@ -2,7 +2,7 @@
 
 namespace App\Exceptions;
 
-use App\Contracts\Exceptions\HttpExceptionHandler;
+use App\Contracts\Exceptions\Handlers\HttpExceptionHandler;
 use App\DTO\BacktraceRecord;
 use App\Exceptions\Handlers\AccessDBDenied;
 use App\Exceptions\Handlers\InstallationHandler;
