+++ app/Exceptions/Handlers/AccessDBDenied.php
@@ -2,8 +2,8 @@
 
 namespace App\Exceptions\Handlers;
 
-use App\Contracts\Exceptions\HttpExceptionHandler;
-use App\Redirections\ToInstall;
+use App\Contracts\Exceptions\Handlers\HttpExceptionHandler;
+use App\Http\Redirections\ToInstall;
 use Illuminate\Database\QueryException;
 use Symfony\Component\HttpFoundation\Response as SymfonyResponse;
 use Symfony\Component\HttpKernel\Exception\HttpExceptionInterface as HttpException;
