+++ app/Http/Redirections/ToMigration.php
@@ -1,19 +1,22 @@
 <?php
 
-namespace App\Redirections;
+namespace App\Http\Redirections;
 
+use App\Contracts\Http\Redirection;
+use Illuminate\Contracts\Container\BindingResolutionException;
 use Illuminate\Http\RedirectResponse;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Exception\RouteNotFoundException;
 
-class ToHome implements Redirection
+class ToMigration implements Redirection
 {
 	/**
 	 * @throws RouteNotFoundException
+	 * @throws BindingResolutionException
 	 */
 	public static function go(): RedirectResponse
 	{
-		return redirect(route('home'), Response::HTTP_TEMPORARY_REDIRECT, [
+		return redirect(route('migrate'), Response::HTTP_TEMPORARY_REDIRECT, [
 			'Cache-Control' => 'no-cache, must-revalidate',
 		]);
 	}
